Refactor: move problem code.ts files to code/typescript.ts
Moved `code.ts` files from individual problem directories under `packages/backend/src/problem/free/` into a `code` subfolder and renamed them to `typescript.ts`.

This standardizes the structure for problem code files.

Updated import paths in affected `problem.ts` files to reflect the new location.
