+++ packages/backend/src/problem/free/houseRobber/code/typescript.ts
